Success screen is implemented when registering a user.
