+++ app/src/main/java/com/deymer/ibank/ui/utils/Utils.kt
@@ -1,9 +1,14 @@
 package com.deymer.ibank.ui.utils
 
+import android.content.Context
+import android.net.Uri
+import android.os.Environment
+import androidx.core.content.FileProvider
 import com.deymer.presentation.R
 import com.deymer.repository.models.TransactionType.DEPOSIT
 import com.deymer.repository.models.TransactionType.WITHDRAWAL
 import com.deymer.repository.models.TransactionType.TRANSFER
+import java.io.File
 
 fun setIconTransaction(
     transactionType: String
@@ -17,3 +22,14 @@ fun setIconTransaction(
 fun transactionWin(
     transactionType: String
 ) = transactionType == DEPOSIT.name
+
+fun createImageUri(context: Context): Uri? {
+    val timeStamp: String = System.currentTimeMillis().toString()
+    val storageDir: File? = context.getExternalFilesDir(Environment.DIRECTORY_PICTURES)
+    val authority = "${context.packageName}.provider"
+    return storageDir?.let {
+        val fileName = "JPEG_${timeStamp}_"
+        val file = File.createTempFile(fileName, ".jpg", it)
+        FileProvider.getUriForFile(context, authority, file)
+    }
+}
